+++ mp_init_size.c
@@ -4,12 +4,12 @@
 /* SPDX-License-Identifier: Unlicense */
 
 /* init an mp_init for a given size */
-mp_err mp_init_size(mp_int *a, int size)
+mp_err mp_init_size(mp_int *a, size_t size)
 {
    size = MP_MAX(MP_MIN_PREC, size);
 
    /* alloc mem */
-   a->dp = (mp_digit *) MP_CALLOC((size_t)size, sizeof(mp_digit));
+   a->dp = (mp_digit *) MP_CALLOC(size, sizeof(mp_digit));
    if (a->dp == NULL) {
       return MP_MEM;
    }
