+++ s_mp_get_bit.c
@@ -5,10 +5,10 @@
 /* SPDX-License-Identifier: Unlicense */
 
 /* Get bit at position b and return true if the bit is 1, false if it is 0 */
-bool s_mp_get_bit(const mp_int *a, unsigned int b)
+bool s_mp_get_bit(const mp_int *a, size_t b)
 {
    mp_digit bit;
-   int limb = (int)(b / MP_DIGIT_BIT);
+   size_t limb = b / MP_DIGIT_BIT;
 
    if (limb >= a->used) {
       return false;
