+++ s_mp_log_pow2.c
@@ -3,10 +3,10 @@
 /* LibTomMath, multiple-precision integer library -- Tom St Denis */
 /* SPDX-License-Identifier: Unlicense */
 
-uint32_t s_mp_log_pow2(const mp_int *a, uint32_t base)
+size_t s_mp_log_pow2(const mp_int *a, uint32_t base)
 {
-   int y;
+   unsigned y;
    for (y = 0; (base & 1u) == 0u; y++, base >>= 1) {}
-   return (uint32_t)((mp_count_bits(a) - 1) / y);
+   return (mp_count_bits(a) - 1) / y;
 }
 #endif
